Refactor agent mediator prompt editing to hide advanced settings in dialog
