+++ frontend/src/components/experiment_builder/agent_base_prompt_editor.scss
@@ -1,0 +1,109 @@
+@use '../../sass/colors';
+@use '../../sass/common';
+@use '../../sass/typescale';
+
+:host {
+  @include common.flex-column;
+  gap: common.$spacing-large;
+  max-width: common.$info-content-max-width;
+  padding: common.$main-content-padding;
+}
+
+.section-header {
+  @include common.flex-row-align-center;
+  flex-wrap: wrap;
+  gap: common.$spacing-medium;
+  justify-content: space-between;
+}
+
+.subsection-header {
+  @include common.flex-column;
+}
+
+.section-title {
+  @include typescale.title-medium;
+}
+
+.small {
+  @include typescale.label-small;
+}
+
+.agent-wrapper {
+  @include common.flex-column;
+  gap: common.$spacing-xxl;
+}
+
+.description {
+  color: var(--md-sys-color-outline);
+  font-style: italic;
+}
+
+.field {
+  @include common.flex-column;
+  gap: common.$spacing-medium;
+}
+
+.number-input {
+  @include common.number-input;
+}
+
+.name-value-input {
+  @include common.flex-row;
+  background: var(--md-sys-color-surface-variant);
+  border-radius: common.$spacing-small;
+  gap: common.$spacing-medium;
+  padding: common.$spacing-medium;
+  width: max-content;
+}
+
+.divider {
+  border-bottom: 1px solid var(--md-sys-color-outline-variant);
+  width: 100%;
+}
+
+.select-field {
+  @include common.flex-column;
+  gap: common.$spacing-small;
+}
+
+select {
+  @include common.select;
+  background: var(--md-sys-color-surface-variant);
+  border-radius: common.$spacing-medium;
+  color: var(--md-sys-color-on-surface-variant);
+}
+
+.button-wrapper {
+  @include common.flex-row-align-center;
+  gap: common.$spacing-small;
+}
+
+.checkbox-wrapper {
+  @include common.flex-row-align-center;
+  gap: common.$spacing-small;
+  overflow-wrap: break-word;
+
+  md-checkbox {
+    flex-shrink: 0;
+  }
+}
+
+.code-wrapper {
+  @include common.flex-column;
+  gap: common.$spacing-small;
+  max-width: calc(common.$info-content-max-width - common.$spacing-xxl);
+  width: 100%;
+}
+
+pre {
+  background: var(--md-sys-color-surface-variant);
+  margin: 0;
+  overflow: auto;
+  padding: common.$spacing-medium;
+}
+
+.section,
+.subsection {
+  @include common.flex-column;
+  gap: common.$spacing-large;
+}
